need login when visit admin pages

+++ lib/src/controllers/admins.dart
@@ -1,5 +1,7 @@
 part of _controllers;
 
+const USER_ID_KEY = "userId";
+
 initData(Request req) {
     Category category = categoryDao.newModel();
     category.name = "Scala";
@@ -25,29 +27,38 @@ initData(Request req) {
 }
 
 adminLogin(Request req) {
-    HttpRequest httpRequest = request.input;
-    if (httpRequest.session.containsKey("userId")) {
-        req.response.send(httpRequest.session["userId"]);
+    HttpRequest httpRequest = req.input;
+    if (httpRequest.session.containsKey(USER_ID_KEY)) {
+        req.response.send(httpRequest.session[USER_ID_KEY]);
+        adminTopics(req);
         return;
     }
     req.response.send(views.login());
 }
 
 adminDoLogin(Request req) {
+    print("### in admiNDoLogin");
     _getPostData(req, (postData) {
         String email = postData['email'], password = postData['password'];
-        var rs = userDao.findBy("email=? and password=?", [email, password]);
-        if (rs.length == 1) {
-            HttpRequest httpRequest = request.input;
-            httpRequest.session.putIfAbsent("userId", () => rs.first.id);
-            req.response.send("ok");
-        } else {
+        var users = userDao.findBy("email=?", [email]);
+        if (users.isEmpty) {
             req.response.send("not found");
+        } else {
+            var user = users.first;
+            if (user.checkPassword(password)) {
+                HttpRequest httpRequest = req.input;
+                httpRequest.session[USER_ID_KEY] = user.id;
+                req.response.send("ok");
+            } else {
+                req.response.send("invalid password");
+            }
         }
     });
 }
 
 writePage(Request req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     var id = req.param("id");
     var categories = categoryDao.listAll("name asc");
     var topic = null;
@@ -58,6 +69,8 @@ writePage(Request req) {
 }
 
 adminSaveTopic(Request req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     _getPostData(req, (postData) {
         String id = postData['id'],
         title = postData['title'],
@@ -84,11 +97,15 @@ adminSaveTopic(Request req) {
 }
 
 adminTopics(Request req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     var topics = topicDao.listAll();
     req.response.send(views.adminTopics(topics));
 }
 
 adminEditTopic(Request req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     var id = req.param("id");
     var topic = topicDao.get(id);
     List categories = categoryDao.listAll("name asc");
@@ -96,28 +113,38 @@ adminEditTopic(Request req) {
 }
 
 adminDeleteTopic(Request req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     var id = req.param("id");
     var topic = topicDao.deleteById(id);
     req.response.send("ok");
 }
 
 adminCategories(Request req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     var categories = categoryDao.listAll("displayOrder desc");
     print("### categories: $categories");
     req.response.send(views.adminCategories(categories));
 }
 
 adminDeleteCategory(req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     var id = req.param("id");
     categoryDao.deleteById(id);
     adminCategories(req);
 }
 
 adminEditCategory(req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     req.response.send("TODO");
 }
 
 adminCreateCategory(req) {
+    if (_isNotLoggedIn(req)) return adminLogin(req);
+
     _getPostData(req, (postData) {
         var name = postData["name"];
         var category = categoryDao.newModel();
@@ -125,4 +152,11 @@ adminCreateCategory(req) {
         category.save();
         adminCategories(req);
     });
+}
+
+bool _isNotLoggedIn(Request req) {
+    HttpRequest httpRequest = req.input;
+    var userId = httpRequest.session[USER_ID_KEY];
+    print("### userId : $userId");
+    return userId == null || userId.trim().isEmpty;
 }
